Revise JOSS paper draft; Close #10; Close #11; Close #13
Revision of the JOSS paper draft. This now includes a proper reference
with a Zenodo DOI for the QMKPy datasets repository (Issue #10).
Furthermore, it briefly explains the need for the QMKPy framework over
existing (commercial) optimization frameworks (Issue #11).
Lastly, it improves the description of the package and highlights that
it primarily functions as a testbed for comparing different solution
algorithms (Issue #13).